organise component directories
