+++ src/components/Pages/About/About.tsx
@@ -3,8 +3,8 @@ import { MDBContainer } from "mdbreact";
 import aboutImage1 from "./assets/samuel-johnsons-house-illustration.jpeg";
 import aboutImage2 from "./assets/whitechapel-gallery-illustration.jpeg";
 import aboutImage3 from "./assets/john-soanes-museum-illustration.jpeg";
-import Section from "../../Section/Section";
 import "./About.css";
+import Section from "../../common/Section/Section";
 
 const About = () => {
