+++ src/components/Pages/Admin/Admin.tsx
@@ -18,6 +18,7 @@ import Credentials from "../../../types/PostRequests/Credentials";
 import "./Admin.css";
 import DisplayForm from "./DisplayForm";
 
+
 const Admin = () => {
 
   const [form, setForm] = useState<string>("");
