+++ src/components/common/Carousel/Carousel.css
