+++ src/components/common/Carousel/Slide.tsx
