+++ src/components/common/Carousel/assets/carousel-illustration-1-2.jpg
