+++ src/components/common/Carousel/assets/carousel-illustration-2-2.jpg
