+++ src/components/common/Carousel/assets/carousel-illustration-3-2.jpg
