+++ src/components/common/Footer/FooterCol.tsx
