+++ src/components/common/LoadingBar/LoadingBar.tsx
