+++ src/components/common/LocatorMap/mapConfig.ts
