+++ src/components/common/SearchBar/SearchBar.tsx
@@ -3,9 +3,9 @@ import { MDBInput } from "mdbreact";
 import { FaSearchLocation } from "react-icons/fa";
 import { connect } from "react-redux";
 import { bindActionCreators, Dispatch } from "redux";
-import RootState from "../../types/State/Root/State";
-import { handleChangeSearch, setRedirect } from "../../actions/actions";
-import * as Actions from "../../types/Actions";
+import RootState from "../../../types/State/Root/State";
+import { handleChangeSearch, setRedirect } from "../../../actions/actions";
+import * as Actions from "../../../types/Actions";
 import "./SeachBar.css";
 
 interface SearchBarProps {
